+++ NEWS.md
@@ -6,6 +6,8 @@
   `deselecteerSoortenKenmerken()` (issue #231)
 * bij `berekenLSVIbasis()` worden de extra kolommen uit `Data_habitat`
   toegevoegd aan de uitvoer van de functie
+* bij duinhabitats (2120 en 2130) worden de bedekkingen berekend t.o.v. het
+  begroeide deel van de duin (issue #240)
 
 # LSVI 0.1.3